// src/app/services/models/leave-dto.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { EmployeeShortDto } from '../models/employee-short-dto';
export interface LeaveDto {
  approvalDate?: string;
  approvedBy?: EmployeeShortDto;
  attachments?: string;
  comments?: string;
  durationDays?: number;
  employee?: EmployeeShortDto;
  endDate?: string;
  halfDay?: boolean;
  id?: number;
  leaveType?: 'ANNUAL_LEAVE' | 'SICK_LEAVE' | 'MATERNITY_LEAVE' | 'PATERNITY_LEAVE' | 'UNPAID_LEAVE' | 'BEREAVEMENT_LEAVE' | 'MARRIAGE_LEAVE' | 'TRAINING_LEAVE' | 'SABBATICAL_LEAVE' | 'OTHER';
  reason?: string;
  rejectionReason?: string;
  startDate?: string;
  status?: 'PENDING' | 'APPROVED' | 'REJECTED' | 'CANCELLED' | 'IN_PROGRESS' | 'COMPLETED';
}